Added RemoteCache object
#Fix in the naming of the config variable to match style for config file.
#Updated the  CachedFileDownloader class to have a RemoteCache object to handle this new feature.
#Fixed test to use new class constructor for CachedFileDownloader

Added the support for in the download, when remote_cache_active is True.
The RemoteCache object is created to handle all the work for remote caching.

RemoteCache will use internally a RemoteCacheBase interface to manage the download of the file and store in the configured endpoint for that RemoteCache, using the protocol internally defined.

(I am not very familiar with python inheritance, but that would be the direction I would follow)